Implement a better handling of plugins
This update includes better handling of plugins which are not named as the
participant's name (such as RandomRain.jar) by implementing a way to get
whichever jar the build server has and renaming it. This also includes
renaming the plugin's name in the plugin.yml for the sanity of the judges.

This update also includes a little bit more information to the judge when
they download the plugin, such as if the plugin did not compile on it's
last commit.

It should be noted that this commit also fixes multiple errors in the
pom.xml and includes a large amount of library shading to ensure the
jenkins library chosen will actually run when the plugin starts.

This commit also removes the command assistance library as it just made
for some messy and unprotected code.